[IMP] tcp: aggregate data on xpl
Before, was done on datasets files.
However, those have non-established format, and it was possible 
to miss a lot of useful data (at least the half for big connection, to 
the full connection for small ones).

Therefore, we read the content of the xpl file, which is much easier
to process than the content of datasets one.